Change IPTC methods to use size_t

+++ include/exiv2/iptc.hpp
@@ -293,11 +293,7 @@ namespace Exiv2 {
           @return 0 if successful;<BR>
                   5 if the binary IPTC data is invalid or corrupt
          */
-        static int decode(
-                  IptcData& iptcData,
-            const byte*     pData,
-                  uint32_t  size
-        );
+        static int decode(IptcData& iptcData, const byte* pData, size_t size);
         /*!
           @brief Encode the IPTC datasets from \em iptcData to a binary
                  representation in IPTC IIM4 format.
